moved color.py from my GD fork to here, fixed the shows trying to natively use RGBW
