+++ shows/__init__.py
@@ -5,3 +5,5 @@ from .one_by_one import OneByOne
 from .random_cells import Random
 from .showbase import ShowBase, load_shows, random_shows
 from .up_down import UpDown
+from .cycle_hsv import CycleHSV
+from .cycle_hsv2 import CycleHSV2
