+++ shows/left_to_right_and_back.py
@@ -1,5 +1,5 @@
 from .showbase import ShowBase
-from color import RGBW
+from color import RGB
 import time
 
 
@@ -30,7 +30,7 @@ class LeftToRightAndBack(ShowBase):
                             r=255
                             g = 0
                             for pix in range(6):
-                                self.tri_grid.set_pixel(cell.get_pixels()[pix], RGBW(r, g, 0, 1), cell.get_id())
+                                self.tri_grid.set_pixel(cell.get_pixels()[pix], RGB(r, g, 0), cell.get_id())
                                 time.sleep(.2)
                                 self.tri_grid.go()
                                 g += 40
@@ -56,7 +56,7 @@ class LeftToRightAndBack(ShowBase):
                             g = 255
                             b = 0
                             for pix in range(6):
-                                self.tri_grid.set_pixel(cell.get_pixels()[5-pix], RGBW(0, g, b, 1), cell.get_id())
+                                self.tri_grid.set_pixel(cell.get_pixels()[5-pix], RGB(0, g, b), cell.get_id())
                                 time.sleep(.2)
                                 self.tri_grid.go()
                                 g -= 40
