+++ shows/up_down.py
@@ -1,5 +1,5 @@
 from .showbase import ShowBase
-from color import RGBW
+from color import RGB
 
 
 class UpDown(ShowBase):
@@ -17,12 +17,12 @@ class UpDown(ShowBase):
                 print('up')
                 for i in self.cells.get_up_cells():
                     print("Up", i.get_id())
-                    self.cells.set_cell_by_cellid(i.get_id(), RGBW(255, 255, 255, 255))
+                    self.cells.set_cell_by_cellid(i.get_id(), RGB( 255, 255, 255))
             else:
                 print('down')
                 for i in self.cells.get_down_cells():
                     print("down", i.get_id())
-                    self.cells.set_cell_by_cellid(i.get_id(), RGBW(255, 0, 0, 0))
+                    self.cells.set_cell_by_cellid(i.get_id(), RGB( 255, 0, 0))
 
             if a == "up":
                 a = "down"
